deleted duplicate files. Tests written to mimic pyfda interface
